// Frontend/Web/src/gen/files_api_pb.ts
export enum UploadFileType {
  /**
   * @generated from enum value: UPLOAD_FILE_TYPE_UNKNOWN = 0;
   */
  UPLOAD_FILE_TYPE_UNKNOWN = 0,

  /**
   * Аватарка пользователя
   *
   * @generated from enum value: USER_AVATAR = 1;
   */
  USER_AVATAR = 1,

  /**
   * Вложение сообщения - изображение
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_IMAGE = 2;
   */
  MESSAGE_ATTACHMENT_IMAGE = 2,

  /**
   * Вложение сообщения - видео
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_VIDEO = 3;
   */
  MESSAGE_ATTACHMENT_VIDEO = 3,

  /**
   * Вложение сообщения - гифка
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_GIF = 4;
   */
  MESSAGE_ATTACHMENT_GIF = 4,

  /**
   * Вложение сообщения - документ
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_DOCUMENT = 5;
   */
  MESSAGE_ATTACHMENT_DOCUMENT = 5,

  /**
   * Обложка группового чата
   *
   * @generated from enum value: CHAT_PICTURE = 6;
   */
  CHAT_PICTURE = 6,

  /**
   * Вложение сообщения - аудио
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_AUDIO = 7;
   */
  MESSAGE_ATTACHMENT_AUDIO = 7,

  /**
   * Вложение сообщения - голосовое сообщение
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_VOICE = 8;
   */
  MESSAGE_ATTACHMENT_VOICE = 8,

  /**
   * Вложение сообщения - стикер (WebP)
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_STICKER = 9;
   */
  MESSAGE_ATTACHMENT_STICKER = 9,

  /**
   * Постер профиля пользователя (горизонтальный баннер)
   *
   * @generated from enum value: USER_PROFILE_POSTER = 10;
   */
  USER_PROFILE_POSTER = 10,
}
// ...
export class GetUserStorageInfoResponse extends Message<GetUserStorageInfoResponse> {
  /**
   * Общее занятое пространство в байтах
   *
   * @generated from field: int64 total_used_storage = 1;
   */
  totalUsedStorage = protoInt64.zero;

  /**
   * Лимит хранилища в байтах
   *
   * @generated from field: int64 storage_limit = 2;
   */
  storageLimit = protoInt64.zero;

  /**
   * Использованное пространство по типам файлов
   *
   * @generated from field: repeated barkfluff.files.GetUserStorageInfoResponse.StorageByType storage_by_types = 3;
   */
  storageByTypes: GetUserStorageInfoResponse_StorageByType[] = [];

  constructor(data?: PartialMessage<GetUserStorageInfoResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "barkfluff.files.GetUserStorageInfoResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "total_used_storage", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 2, name: "storage_limit", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 3, name: "storage_by_types", kind: "message", T: GetUserStorageInfoResponse_StorageByType, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetUserStorageInfoResponse {
    return new GetUserStorageInfoResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): GetUserStorageInfoResponse {
    return new GetUserStorageInfoResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): GetUserStorageInfoResponse {
    return new GetUserStorageInfoResponse().fromJsonString(jsonString, options);
  }

  static equals(a: GetUserStorageInfoResponse | PlainMessage<GetUserStorageInfoResponse> | undefined, b: GetUserStorageInfoResponse | PlainMessage<GetUserStorageInfoResponse> | undefined): boolean {
    return proto3.util.equals(GetUserStorageInfoResponse, a, b);
  }
}

/**
 * @generated from message barkfluff.files.GetUserStorageInfoResponse.StorageByType
 */
export class GetUserStorageInfoResponse_StorageByType extends Message<GetUserStorageInfoResponse_StorageByType> {
  /**
   * Тип файла
   *
   * @generated from field: barkfluff.files.UploadFileType file_type = 1;
   */
  fileType = UploadFileType.UPLOAD_FILE_TYPE_UNKNOWN;

  /**
   * Занятое пространство для данного типа в байтах
   *
   * @generated from field: int64 used_storage = 2;
   */
  usedStorage = protoInt64.zero;

  constructor(data?: PartialMessage<GetUserStorageInfoResponse_StorageByType>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "barkfluff.files.GetUserStorageInfoResponse.StorageByType";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "file_type", kind: "enum", T: proto3.getEnumType(UploadFileType) },
    { no: 2, name: "used_storage", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetUserStorageInfoResponse_StorageByType {
    return new GetUserStorageInfoResponse_StorageByType().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): GetUserStorageInfoResponse_StorageByType {
    return new GetUserStorageInfoResponse_StorageByType().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): GetUserStorageInfoResponse_StorageByType {
    return new GetUserStorageInfoResponse_StorageByType().fromJsonString(jsonString, options);
  }

  static equals(a: GetUserStorageInfoResponse_StorageByType | PlainMessage<GetUserStorageInfoResponse_StorageByType> | undefined, b: GetUserStorageInfoResponse_StorageByType | PlainMessage<GetUserStorageInfoResponse_StorageByType> | undefined): boolean {
    return proto3.util.equals(GetUserStorageInfoResponse_StorageByType, a, b);
  }
}